Fix date parse to local on Reactnative JS, add created_at in user timezone on user practice
